requirements.txt
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
frozenlist==1.8.0
google-auth==2.49.1
google-auth-oauthlib==1.3.0
greenlet==3.3.2
gspread==6.2.1
idna==3.11
magic-filter==1.0.12
multidict==6.7.1
oauthlib==3.3.1
propcache==0.4.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
SQLAlchemy==2.0.48
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
